Crate zng_wgt_wizard

Wizard widget.

Crate

This crate is part of the zng project.

Structs

ContentFnArgs

Arguments for a wizard main content container builder.

FooterFnArgs

Arguments for a wizard footer container builder.

HeaderFnArgs

Arguments for a wizard header container builder.

Page

Represents a page builder for Wizard!.

PageArgs

Arguments for Page builders.

PanelFnArgs

Arguments for the wizard root panel builder.

SideFnArgs

Arguments for a wizard side container builder.

Wizard

W Paginated widget that

Statics

BACK_CMD

Return to previous page.

CANCEL_CMD

Cancel wizard operation.

CAN_CANCEL_VAR

Defines if on_cancel and on_pre_cancel command handlers are enabled in a widget and descendants.

CAN_FINISH_VAR

Defines if on_finish and on_pre_finish command handlers are enabled in a widget and descendants.

CONTENT_FN_VAR

Widget function that builds a wizard main content container.

FINISH_CMD

Finish wizard operation.

FOOTER_EXTRA_FN_VAR

Widget function that builds extra content for the footer panel.

FOOTER_FN_VAR

Widget function that builds a wizard footer container.

HEADER_BACKGROUND_FN_VAR

Widget function that builds a background visual for the header container.

HEADER_FN_VAR

Widget function that builds a wizard header container.

NEXT_CMD

Advance to next page.

PANEL_FN_VAR

Widget function that brings together all the wizard parts.

SIDE_BACKGROUND_FN_VAR

Widget function that builds a background visual for the side container.

SIDE_EXTRA_FN_VAR

Widget function that builds extra content for the side panel for pages that define side content.

SIDE_FN_VAR

Widget function that builds a wizard side container.

Functions

can_cancel

P Defines if on_cancel and on_pre_cancel command handlers are enabled in the widget and descendants.

can_finish

P Defines if on_finish and on_pre_finish command handlers are enabled in the widget and descendants.

content_fn

P Widget function that converts a ContentFnArgs into a page main content container container widget.

default_content_fn

Default wizard main content container.

default_footer_fn

Default wizard footer container.

default_header_fn

Default wizard header container.

default_page_footer

Default wizard page footer content.

default_page_footer_back

Default BACK_CMD button.

default_page_footer_cancel

Default CANCEL_CMD button.

default_page_footer_finish

Default FINISH_CMD button.

default_page_footer_next

Default NEXT_CMD button.

default_page_header

Default wizard page header content.

default_page_header_info

Default Page::info presenter.

default_page_header_title

Default Page::title presenter.

default_page_side

Default wizard page side content.

default_panel_fn

Default wizard root panel.

default_side_fn

Default wizard side container.

finish_cmd_name

P Set the name for the FINISH_CMD scoped on this widget.

footer_extra_fn

P Widget function that converts PageArgs to extra footer content.

footer_fn

P Widget function that converts a FooterFnArgs into a page side container widget.

get_title

P Get the current page title.

header_background_fn

P Widget function that makes a background visual for the page header container widget.

header_fn

P Widget function that converts a HeaderFnArgs into a page header container widget.

on_cancel

P Wizard cancel requested.

on_finish

P Wizard finish requested.

on_pre_cancel

P Wizard cancel requested.

on_pre_finish

P Wizard finish requested.

pages

B Defines the wizard pages.

panel_fn

P Widget function that converts a PanelFnArgs into a wizard.

side_background_fn

P Widget function that makes a background visual for the page side container widget.

side_extra_fn

P Widget function that converts PageArgs to extra side panel content.

side_fn

P Widget function that converts a SideFnArgs into a page side container widget.